refactor: rearrange dir
