+++ src/main/kotlin/ecommerce/config/argumentResolver/LoginMemberArgumentResolver.kt
@@ -1,10 +1,10 @@
 package ecommerce.config.argumentResolver
 
-import ecommerce.annotations.LoginMember
-import ecommerce.entity.User
 import ecommerce.enums.UserRole
-import ecommerce.exception.UnauthorisedUserException
+import ecommerce.model.User
 import ecommerce.repository.UserRepository
+import ecommerce.utils.annotations.LoginMember
+import ecommerce.utils.exception.UnauthorisedUserException
 import org.springframework.core.MethodParameter
 import org.springframework.web.bind.support.WebDataBinderFactory
 import org.springframework.web.context.request.NativeWebRequest
