+++ src/main/kotlin/ecommerce/config/interceptor/AdminInterceptor.kt
@@ -1,10 +1,10 @@
 package ecommerce.config.interceptor
 
-import ecommerce.entity.User
 import ecommerce.enums.UserRole
-import ecommerce.exception.UnauthorisedUserException
-import ecommerce.infrastructure.JwtProvider
+import ecommerce.model.User
 import ecommerce.repository.UserRepository
+import ecommerce.utils.exception.UnauthorisedUserException
+import ecommerce.utils.infrastructure.JwtProvider
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
 import org.springframework.stereotype.Component
