+++ src/main/kotlin/ecommerce/config/interceptor/BaseAuthInterceptor.kt
@@ -1,9 +1,9 @@
 package ecommerce.config.interceptor
 
-import ecommerce.entity.User
-import ecommerce.exception.UnauthorisedUserException
-import ecommerce.infrastructure.JwtProvider
+import ecommerce.model.User
 import ecommerce.repository.UserRepository
+import ecommerce.utils.exception.UnauthorisedUserException
+import ecommerce.utils.infrastructure.JwtProvider
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
 import org.springframework.web.servlet.HandlerInterceptor
