+++ src/main/kotlin/ecommerce/model/CartStatistics.kt
@@ -1,4 +1,4 @@
-package ecommerce.entity
+package ecommerce.model
 
 import ecommerce.enums.CartAction
 import jakarta.persistence.Column
