+++ src/main/kotlin/ecommerce/model/Product.kt
@@ -1,4 +1,4 @@
-package ecommerce.entity
+package ecommerce.model
 
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
