+++ src/main/kotlin/ecommerce/repository/CartProductRepository.kt
@@ -1,6 +1,6 @@
 package ecommerce.repository
 
-import ecommerce.entity.CartProduct
+import ecommerce.model.CartProduct
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
