+++ src/main/kotlin/ecommerce/repository/CartRepository.kt
@@ -1,6 +1,6 @@
 package ecommerce.repository
 
-import ecommerce.entity.Cart
+import ecommerce.model.Cart
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
