+++ src/main/kotlin/ecommerce/repository/ProductRepository.kt
@@ -1,6 +1,6 @@
 package ecommerce.repository
 
-import ecommerce.entity.Product
+import ecommerce.model.Product
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 import java.util.Optional
