+++ src/main/kotlin/ecommerce/repository/UserRepository.kt
@@ -1,6 +1,6 @@
 package ecommerce.repository
 
-import ecommerce.entity.User
+import ecommerce.model.User
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 import java.util.Optional
