+++ src/main/kotlin/ecommerce/service/AdminProductService.kt
@@ -3,10 +3,10 @@ package ecommerce.service
 import ecommerce.dto.products.ProductDTO
 import ecommerce.dto.products.ProductPatchDTO
 import ecommerce.dto.products.ProductResponseDTO
-import ecommerce.entity.Product
-import ecommerce.exception.DuplicateProductNameException
-import ecommerce.exception.EntityNotFoundException
+import ecommerce.model.Product
 import ecommerce.repository.ProductRepository
+import ecommerce.utils.exception.DuplicateProductNameException
+import ecommerce.utils.exception.EntityNotFoundException
 import ecommerce.utils.toProductDTO
 import org.springframework.stereotype.Service
 import java.net.URI
