+++ src/main/kotlin/ecommerce/utils/exception/CartOperationException.kt
@@ -1,3 +1,3 @@
-package ecommerce.exception
+package ecommerce.utils.exception
 
 class CartOperationException(message: String) : RuntimeException(message)
