+++ src/main/kotlin/ecommerce/utils/exception/DuplicateProductNameException.kt
@@ -1,3 +1,3 @@
-package ecommerce.exception
+package ecommerce.utils.exception
 
 class DuplicateProductNameException(name: String) : RuntimeException("Product with name '$name' already exists.")
