+++ src/main/kotlin/ecommerce/utils/exception/EntityNotFoundException.kt
@@ -1,3 +1,3 @@
-package ecommerce.exception
+package ecommerce.utils.exception
 
 class EntityNotFoundException(error: String) : RuntimeException(error)
