+++ src/main/kotlin/ecommerce/utils/exception/UnauthorisedUserException.kt
@@ -1,3 +1,3 @@
-package ecommerce.exception
+package ecommerce.utils.exception
 
 class UnauthorisedUserException(message: String? = null) : RuntimeException(message ?: "Please log in first")
