+++ src/main/kotlin/ecommerce/utils/exception/UserCredentialException.kt
@@ -1,3 +1,3 @@
-package ecommerce.exception
+package ecommerce.utils.exception
 
 class UserCredentialException(message: String? = null) : RuntimeException(message ?: "Not valid Credential")
