+++ src/main/resources/application.properties
@@ -10,3 +10,4 @@ security.jwt.token.expire-length=${JWT_TIME}
 spring.jpa.hibernate.ddl-auto=create
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
+#spring.jpa.open-in-view=false
